Replace deprecated OpenStruct with MethodAccessibleHash
Replace the deprecated OpenStruct with a custom MethodAccessibleHash
class to ensure compatibility with Ruby 3.5+. OpenStruct is deprecated
as of Ruby 3.4 and will be removed in future versions.

Add MethodAccessibleHash class that provides the same functionality as
OpenStruct but uses 'method_missing' instead of defining methods
on-the-fly. Update 'osparams' method to use MethodAccessibleHash.

Remove 'require ostruct' from lib/inquery.rb and add corresponding
tests for MethodAccessibleHash to ensure proper functionality.

Also fix typo in comment: 'easyer' -> 'convenient'.